+++ callautomation-az-openai-voice/package.json
@@ -17,6 +17,7 @@
     "@azure/openai": "^2.0.0-beta.2",
     "@types/express": "^4.17.17",
     "@types/node": "^20.2.1",
+    "callautomation_az_openai_voice": "file:",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "openai": "^4.72.0",
