+++ callautomation-live-transcription/package.json
@@ -5,8 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "tsc",
-    "dev:app": "nodemon ./src/app.ts",
-    "dev:websocket": "nodemon ./src/websocket.ts"
+    "dev": "nodemon ./src/app.ts"
   },
   "keywords": [],
   "author": "",
@@ -18,6 +17,7 @@
     "@azure/openai": "^1.0.0-beta.7",
     "@types/express": "^4.17.17",
     "@types/node": "^20.2.1",
+    "callautomation_live_transcription": "file:",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "uuid": "^9.0.1",
